requirements.txt
wxPython
pyserial
